1. Fixed a bug in the filewriter that caused the files to become empty if existed. 2. Changed the xbeans extension to write the Databinding classes inside the given package. 3. Changed the security module to suit the changes in the code generation

+++ modules/wsdl/src/org/apache/axis2/wsdl/codegen/writer/ClassWriter.java
@@ -114,7 +114,9 @@ public abstract class ClassWriter {
                 getFileExtensionForLanguage(language));
         //set the existing flag
         fileExists = outputFile.exists();
-        this.stream = new FileOutputStream(outputFile);
+        if (!fileExists){
+            this.stream = new FileOutputStream(outputFile);
+        }
     }
 
     /**
