+++ modules/wsdl/src/org/apache/axis2/wsdl/codegen/writer/ServiceXMLWriter.java
@@ -40,9 +40,11 @@ public class ServiceXMLWriter extends ClassWriter {
                 packageName,
                 "services",
                 ".xml");
-         //set the existing flag
+        //set the existing flag
         fileExists = outputFile.exists();
-        this.stream = new FileOutputStream(outputFile);
+        if (!fileExists){
+            this.stream = new FileOutputStream(outputFile);
+        }
     }
 }
