+++ modules/wsdl/src/org/apache/axis2/wsdl/codegen/writer/TestServiceXMLWriter.java
@@ -36,8 +36,10 @@ public class TestServiceXMLWriter extends ServiceXMLWriter {
                 packageName,
                 "testservice",
                 ".xml");
-         //set the existing flag
+        //set the existing flag
         fileExists = outputFile.exists();
-        this.stream = new FileOutputStream(outputFile);
+        if (!fileExists){
+            this.stream = new FileOutputStream(outputFile);
+        }
     }
 }
